feat(optimizer@m-3000646081): open ansi pipeline with optimzier

See merge request: !23295

+++ tests/queries/4_cnch_stateless/02676_to_decimal_string_ck233.sql
@@ -4,9 +4,9 @@ SELECT toDecimalString(2.123456, 2);  -- rounding
 SELECT toDecimalString(-2, 77);  -- more digits required than exist
 SELECT toDecimalString(-2.123456, 2);  -- rounding
 
-SELECT toDecimalString(2.9876, 60);  -- more digits required than exist (took 60 as it is float by default)
+SELECT toDecimalString(2.9876::Float64, 60);  -- more digits required than exist (took 60 as it is float by default)
 SELECT toDecimalString(2.1456, 2);  -- rounding
-SELECT toDecimalString(-2.9876, 60);  -- more digits required than exist
+SELECT toDecimalString(-2.9876::Float64, 60);  -- more digits required than exist
 SELECT toDecimalString(-2.1456, 2);  -- rounding
 
 -- Float32 and Float64 tests. No sense to test big float precision -- the result will be a mess anyway.
